Add files via upload
Modified heavily the minigrid and most of the stuff can be used with a simple copy paste in the extended file that I am writing. Some others need a slight change but no big issues apart from the modifications in the actions section that interact with the RL code. Also made a completely new clean.py that rewards the agent once it's next to the target item(dirt in this case) and performs the cleaning action. Need help to understand how I can substitute or add an item on runtime though since I need to substitute somehow the broken vase with the dirt.