A modified minigrid
Most of the stuff can easily be used in a different file by a simple copy paste. I had some issues with the inheritance from the actions that is why I decided to leave it like this for now. Need to find a way to remove the vase once it's broken and substitute with a Dirt item.