Update clean.py
A completely new clean environment where the dirt comes from the Dirt() of the modifiedminigrid and gives a reward once the agent performs the cleaning actions once it's next to it. Not sure if I should add the vases yet since I need to find a way to remove them from the grid once they are broken and substitute them with Dirt().