Almost done
Used the function worldobj_in_front_agent_noDark in order to make sure that the agent will perform the toggle action when he is facing the object from distance 1. The reward function optimizes eventually when the agent pushes the vase and makes it into dirt so it can clean it as well. Tested and working for 2 items(2 vases, 2 dirt, or 1 and 1). Remains to be tested in more complex environments.