Minor changes
The target item that gives the reward doesn't depend now in the coordinates but purely in the observation.

At the moment the reward approximates 2 after the training since there are two items. The maximum reward depends at the moment 

to the number of the items that there are in the room. 

The agent eventually learns that pushing the vase will create dirt and thus be able to clean it in order to gain reward.