Unsafe Reward Hacking Env
Same as the rewardhacking.py but added a list of coordinates from while placing the dirt initially. During the step if the toggle action is performed on the dirt it removes the coordinates from list. If the toggle is performed on a vase it adds it's coordinates on the list. 
TODO: Don't take the coordinates of the dirt initially but rather from  the agents scanning action(To be implemented as part of the extendedminigrid).